feat(ai-usage,chatbot,feedback,installers,super-admin): refine event tracking and data handling
- Remove unnecessary type casting (::int) from token sum aggregations in AI usage queries
- Refactor chatbot agent streaming to log one event per ReAct reasoning step instead of only final cumulative usage
- Update agent to track currentNodeUsage per model_request node and flush on node boundaries
- Fix card generation to properly extract usage_metadata from raw response when using includeRaw: true
- Improve feedback date filtering with proper UTC end-of-day boundary handling for inclusive date ranges
- Add comprehensive comments explaining streaming node boundaries and usage metadata extraction patterns
- Update DTOs with improved documentation and validation for installer profiles and assignments
- Enhance super-admin controller with clearer response handling and error management

+++ src/modules/ai-usage/actions/ai-usage-event.action.ts
@@ -78,8 +78,8 @@ export class AiUsageEventModelAction extends AbstractModelAction<AiUsageEvent> {
     const sql = `
       SELECT
         TO_CHAR(DATE_TRUNC('${trunc}', e.created_at), ${labelFmt}) AS label,
-        COALESCE(SUM(e.input_tokens), 0)::int                      AS "inputTokens",
-        COALESCE(SUM(e.output_tokens), 0)::int                     AS "outputTokens"
+        COALESCE(SUM(e.input_tokens), 0)                      AS "inputTokens",
+        COALESCE(SUM(e.output_tokens), 0)                     AS "outputTokens"
       FROM ai_usage_events e
       WHERE
         e.deleted_at IS NULL
